requirements.txt
python-binance
TA-Lib
numpy
backtrader
Flask
